backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
gradio_client>=1.4.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
msal>=1.28.0
google-api-python-client>=2.140.0
google-auth>=2.29.0
dropbox>=11.36.0
boto3>=1.34.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.14.0
